Add files via upload
Asseto Corsa app/mod
Version 1.0

FCW Front Collision Warning is a vehicle safety feature that uses sensors to detect the distance between your car and the vehicle in front of you. If the system determines that a collision is likely, it will alert the driver, through visual and audible warnings. FCW system do not automatically brake the vehicle.